requirements.txt
crewai
langchain
ollama
python-dotenv
langchain-ollama
pydantic
pydantic-settings
PyYAML
